@@ -8,8 +8,8 @@
     "@codemirror/view": "^6.34.1",
     "@emoji-mart/data": "^1.2.1",
     "@emoji-mart/react": "^1.1.1",
-    "@gatsbyjs/reach-router": "1.3.9",
-    "@seafile/react-image-lightbox": "3.0.4",
+    "@gatsbyjs/reach-router": "2.0.1",
+    "@seafile/react-image-lightbox": "4.0.0",
     "@seafile/resumablejs": "1.1.16",
     "@seafile/sdoc-editor": "1.0.217",
     "@seafile/seafile-calendar": "0.0.28",
@@ -36,13 +36,13 @@
     "object-assign": "4.1.1",
     "prop-types": "^15.8.1",
     "qrcode.react": "^1.0.1",
-    "react": "17.0.2",
+    "react": "18.3.1",
     "react-app-polyfill": "^2.0.0",
     "react-chartjs-2": "4.0.0",
     "react-cookies": "^0.1.0",
     "react-dnd": "^2.6.0",
     "react-dnd-html5-backend": "^2.6.0",
-    "react-dom": "17.0.2",
+    "react-dom": "18.3.1",
     "react-i18next": "^10.12.2",
     "react-responsive": "9.0.2",
     "react-select": "5.7.0",
@@ -63,7 +63,7 @@
     "start": "node scripts/start.js",
     "build": "node scripts/build.js",
     "test": "node scripts/test.js --env=jsdom",
-    "dev": "export NODE_ENV=development && node config/server.js"
+    "dev": "export NODE_ENV=development && node --max-old-space-size=4096 config/server.js"
   },
   "browserslist": {
     "production": [
