@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import FileView from './components/file-view/file-view';
 import FileViewTip from './components/file-view/file-view-tip';
 import Image from './components/file-content-view/image';
@@ -88,4 +88,5 @@ class InnerFileView extends React.Component {
   }
 }
 
-ReactDom.render(<InnerFileView />, document.getElementById('wrapper'));
+const root = createRoot(document.getElementById('wrapper'));
+root.render(<InnerFileView />);
