@@ -1,6 +1,6 @@
 // Import React!
 import React, { Suspense } from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 import i18n from './_i18n/i18n-seafile-editor';
 import MarkdownEditor from './pages/markdown-editor';
@@ -12,7 +12,8 @@ import './index.css';
 
 const { repoID, repoEncrypted, filePerm } = window.app.pageOptions;
 
-ReactDom.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <I18nextProvider i18n={ i18n } >
     <Suspense fallback={<Loading />}>
       <MetadataStatusProvider repoID={repoID} repoInfo={{ permission: filePerm, encrypted: repoEncrypted }}>
@@ -21,6 +22,5 @@ ReactDom.render(
         </CollaboratorsProvider>
       </MetadataStatusProvider>
     </Suspense>
-  </I18nextProvider>,
-  document.getElementById('root')
+  </I18nextProvider>
 );
