@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 import i18n from '../../../_i18n/i18n-sdoc-editor';
 import { UncontrolledTooltip } from 'reactstrap';
@@ -436,8 +436,9 @@ class SdocFileHistory extends React.Component {
   }
 }
 
-ReactDom.render(
+const root = createRoot(document.getElementById('wrapper'));
+root.render(
   <I18nextProvider i18n={ i18n } >
     <SdocFileHistory />
-  </I18nextProvider>,
-  document.getElementById('wrapper'));
+  </I18nextProvider>
+);
