@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 import { PublishedRevisionViewer } from '@seafile/sdoc-editor';
 import i18n from '../../../_i18n/i18n-sdoc-editor';
@@ -45,11 +45,11 @@ window.seafile = {
   revisionId,
 };
 
-ReactDom.render(
+const root = createRoot(document.getElementById('wrapper'));
+root.render(
   <I18nextProvider i18n={ i18n } >
     <Suspense fallback={<Loading />}>
       <PublishedRevisionViewer />
     </Suspense>
-  </I18nextProvider>,
-  document.getElementById('wrapper')
+  </I18nextProvider>
 );
