@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import PropTypes from 'prop-types';
 import { navigate } from '@gatsbyjs/reach-router';
 import dayjs from 'dayjs';
@@ -345,4 +345,5 @@ Item.propTypes = {
   item: PropTypes.object.isRequired,
 };
 
-ReactDom.render(<RepoHistory />, document.getElementById('wrapper'));
+const root = createRoot(document.getElementById('wrapper'));
+root.render(<RepoHistory />);
