@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import MD5 from 'MD5';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { Button, Dropdown, DropdownToggle, DropdownItem, UncontrolledTooltip } from 'reactstrap';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
@@ -1015,4 +1015,5 @@ GridItem.propTypes = {
   showImagePopup: PropTypes.func,
 };
 
-ReactDom.render(<SharedDirView />, document.getElementById('wrapper'));
+const root = createRoot(document.getElementById('wrapper'));
+root.render(<SharedDirView />);
