@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import SharedFileView from './components/shared-file-view/shared-file-view';
 import SharedFileViewTip from './components/shared-file-view/shared-file-view-tip';
 import AudioPlayer from './components/audio-player';
@@ -38,4 +38,5 @@ class FileContent extends React.Component {
   }
 }
 
-ReactDom.render(<SharedFileViewAudio />, document.getElementById('wrapper'));
+const root = createRoot(document.getElementById('wrapper'));
+root.render(<SharedFileViewAudio />);
