@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { SimpleViewer } from '@seafile/sdoc-editor';
 import { I18nextProvider } from 'react-i18next';
 import i18n from './_i18n/i18n-sdoc-editor';
@@ -44,11 +44,12 @@ window.seafile = {
   document.getElementById('favicon').href = fileIcon;
 })();
 
-ReactDom.render(
+
+const root = createRoot(document.getElementById('wrapper'));
+root.render(
   <I18nextProvider i18n={ i18n } >
     <Suspense fallback={<Loading />}>
       <SimpleViewer />
     </Suspense>
-  </I18nextProvider>,
-  document.getElementById('wrapper')
+  </I18nextProvider>
 );
