@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import SharedFileView from './components/shared-file-view/shared-file-view';
 import SharedFileViewTip from './components/shared-file-view/shared-file-view-tip';
 import SeafileCodeMirror from './components/seafile-codemirror';
@@ -27,4 +27,5 @@ class SharedFileViewText extends React.Component {
   }
 }
 
-ReactDom.render(<SharedFileViewText />, document.getElementById('wrapper'));
+const root = createRoot(document.getElementById('wrapper'));
+root.render(<SharedFileViewText />);
