@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import Logo from './components/logo';
 import Account from './components/common/account';
 import TermsPreviewWidget from './components/terms-preview-widget';
@@ -33,4 +33,5 @@ class TCView extends React.Component {
   }
 }
 
-ReactDom.render(<TCView />, document.getElementById('wrapper'));
+const root = createRoot(document.getElementById('wrapper'));
+root.render(<TCView />);
