@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import ReactDom from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 import i18n from './_i18n/i18n-sdoc-editor';
 import { Utils } from './utils/utils';
@@ -59,7 +59,8 @@ window.seafile = {
 
 const repoInfo = { encrypted: repoEncrypted, permission: filePerm, is_admin: isRepoAdmin };
 
-ReactDom.render(
+const root = createRoot(document.getElementById('wrapper'));
+root.render(
   <I18nextProvider i18n={ i18n } >
     <Suspense fallback={<Loading />}>
       <MetadataStatusProvider repoID={repoID} repoInfo={repoInfo}>
@@ -70,6 +71,5 @@ ReactDom.render(
         </CollaboratorsProvider>
       </MetadataStatusProvider>
     </Suspense>
-  </I18nextProvider>,
-  document.getElementById('wrapper')
+  </I18nextProvider>
 );
